extracted component to use in other areas of the application
