+++ src/features/subclass/SubclassCustomizationWrapper.tsx
@@ -2,6 +2,7 @@ import React from 'react';
 import { useSelector } from 'react-redux';
 import { createSelector } from '@reduxjs/toolkit';
 import AbilitiesModification from './AbilitiesModification';
+import StatModifications from './StatModifications';
 import './SubclassCustomizationWrapper.css';
 import { Button, Box } from '@mui/material';
 import { styled } from '@mui/material/styles';
@@ -27,26 +28,6 @@ const TransparentButton = styled(Button)(({ theme }) => ({
   },
 }));
 
-const StatModification = styled('div')(({ theme }) => ({
-  fontFamily: 'Helvetica, Arial, sans-serif',
-  fontWeight: 'bold',
-  fontSize: '1rem',
-  marginBottom: theme.spacing(0.5),
-  textShadow: '1px 1px 1px rgba(0, 0, 0, 0.5)',
-}));
-
-const StatModificationsContainer = styled(Box)(({ theme }) => ({
-  position: 'absolute',
-  bottom: 16,
-  left: 16,
-  zIndex: 1000,
-  padding: theme.spacing(2),
-  backgroundColor: 'rgba(0, 0, 0, 0.2)',
-  backdropFilter: 'blur(5px)',
-  borderRadius: 0,
-  border: '1px solid rgba(255, 255, 255, 0.3)',
-}));
-
 const selectFragmentStatModifications = createSelector(
   (state: RootState) => state.loadoutConfig.loadout.subclassConfig.fragments,
   (fragments: ManifestStatPlug[]) =>
@@ -56,13 +37,12 @@ const selectFragmentStatModifications = createSelector(
         stats.forEach((stat) => {
           const value = fragment[`${stat}Mod` as keyof ManifestStatPlug] as number;
           if (value !== 0) {
-            if (!acc[stat]) acc[stat] = [];
-            acc[stat].push({ name: fragment.name, value });
+            acc.push({ stat, value, name: fragment.name });
           }
         });
       }
       return acc;
-    }, {} as Record<string, { name: string; value: number }[]>)
+    }, [] as { stat: string; value: number; name: string }[])
 );
 
 const SubclassCustomizationWrapper: React.FC<SubclassCustomizationWrapperProps> = ({
@@ -71,21 +51,6 @@ const SubclassCustomizationWrapper: React.FC<SubclassCustomizationWrapperProps> 
   screenshot,
 }) => {
   const fragmentStatMods = useSelector(selectFragmentStatModifications);
-
-  const renderStatModifications = () => {
-    return Object.entries(fragmentStatMods).map(([stat, modifications]) => {
-      return modifications.map(({ name, value }, index) => {
-        const color = value > 0 ? 'green' : 'red';
-        const sign = value > 0 ? '+' : '';
-        return (
-          <StatModification key={`${stat}-${index}`} style={{ color }}>
-            {stat.charAt(0).toUpperCase() + stat.slice(1)}: {sign}
-            {value} ({name})
-          </StatModification>
-        );
-      });
-    });
-  };
 
   return (
     <Box
@@ -100,7 +65,7 @@ const SubclassCustomizationWrapper: React.FC<SubclassCustomizationWrapperProps> 
           Back
         </TransparentButton>
       </Box>
-      <StatModificationsContainer>{renderStatModifications()}</StatModificationsContainer>
+      <StatModifications modifications={fragmentStatMods} />
       <Box sx={{ marginTop: '15vh' }}>
         <AbilitiesModification subclass={subclass} />
       </Box>
